Introduced the extraction and the storage of doxygen-based description for C functions.
This is realized using the two new adapters to maintain API-backward compability.

The first adapter is introducd in ASTTranslationUnit class (IASTTranslationUnit interface). It can
be used to get access to the doxygen map, which maps IASTNodes to plain description strings. This is
actually used by the CFunction to find a description of the entidy.

The second new adapter is currently returned CFunction classes (the plain one and the PDOM one). It
provides access to the description of the function in a backward-compatible manner. In the PDOM, the
description string is stored as an additional record for the function.

The whole mechansim may be generalized to other types of bindings in later commits.

Signed-off-by: Sebastian Bauer <mail@sebastianbauer.info>